requirements.txt
discord.py
asyncio
ffmpeg-python
PyNaCl
python-dotenv